ads_app/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
google-genai==0.3.0
google-cloud-bigquery==3.14.1
requests==2.31.0
